requirements.txt
google-genai>=0.2.0
pyreadline3>=3.4.1; platform_system == "Windows"
